@@ -5,14 +5,14 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>fr.cel</groupId>
-    <artifactId>Parkour</artifactId>
-    <version>1.0</version>
+    <artifactId>parkour</artifactId>
+    <version>1.0.1</version>
     <packaging>jar</packaging>
 
     <name>Parkour</name>
 
     <properties>
-        <java.version>17</java.version>
+        <java.version>21</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
@@ -40,21 +40,21 @@
         <dependency>
             <groupId>org.spigotmc</groupId>
             <artifactId>spigot-api</artifactId>
-            <version>1.20.1-R0.1-SNAPSHOT</version>
+            <version>1.20.6-R0.1-SNAPSHOT</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>1.18.32</version>
             <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>fr.cel</groupId>
             <artifactId>api-plugin</artifactId>
-            <version>1.0</version>
+            <version>1.0.2</version>
             <scope>provided</scope>
         </dependency>
-        <dependency>
-	    <groupId>org.projectlombok</groupId>
-	    <artifactId>lombok</artifactId>
-	    <version>1.18.30</version>
-	    <scope>provided</scope>
-	</dependency>
     </dependencies>
     
 </project>
