@@ -1,21 +1,26 @@
 package fr.cel.cachecache;
 
 import fr.cel.cachecache.commands.CCCommand;
+import fr.cel.cachecache.manager.CCArenaManager;
+import fr.cel.cachecache.manager.GameManager;
 import fr.cel.gameapi.GameAPI;
 import lombok.Getter;
+import lombok.Setter;
 import org.bukkit.plugin.java.JavaPlugin;
 
-import fr.cel.cachecache.manager.CCGameManager;
-
+@Getter
 public class CacheCache extends JavaPlugin {
 
-    @Getter private CCGameManager gameManager;
+    private GameManager gameManager;
+    @Setter private CCArenaManager ccArenaManager;
 
     @Override
     public void onEnable() {
-        this.gameManager = new CCGameManager(this);
+        this.gameManager = new GameManager(this);
+        this.ccArenaManager = new CCArenaManager(this);
+        gameManager.reloadTemporaryHub();
 
-        GameAPI.getInstance().getCommandsManager().addCommand(getCommand("cachecache"), new CCCommand(gameManager));
+        GameAPI.getInstance().getCommandsManager().addCommand("cachecache", new CCCommand(gameManager),  this);
     }
 
     @Override
