@@ -1,19 +1,19 @@
 package fr.cel.cachecache.utils;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-
+import fr.cel.cachecache.CacheCache;
+import fr.cel.cachecache.manager.GameManager;
+import fr.cel.cachecache.manager.GroundItem;
+import fr.cel.cachecache.arena.CCArena;
+import fr.cel.gameapi.utils.LocationUtility;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.configuration.file.YamlConfiguration;
 
-import fr.cel.cachecache.CacheCache;
-import fr.cel.cachecache.manager.arena.CCArena;
-import fr.cel.cachecache.manager.CCGameManager;
-import fr.cel.cachecache.manager.GroundItem;
+import java.io.File;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
 public class Config {
 
@@ -38,9 +38,10 @@ public class Config {
                         arenaName,
                         config.getString("displayName"),
                         CCArena.HunterMode.valueOf(config.getString("hunterMode")),
-                        parseStringToLoc(config.getString("locationSpawn")),
-                        parseStringToLoc(config.getString("locationWaiting")),
-                        config.getBoolean("fallDamage")
+                        LocationUtility.parseConfigToLoc(config, "locationSpawn"),
+                        LocationUtility.parseConfigToLoc(config, "locationWaiting"),
+                        config.getBoolean("fallDamage"),
+                        main.getGameManager()
                 );
 
                 arena.setConfig(this);
@@ -69,7 +70,7 @@ public class Config {
     }
 
     private GroundItem getItemByName(String itemName) {
-        for (GroundItem item : CCGameManager.getGroundItems()) {
+        for (GroundItem item : main.getGameManager().getGroundItems()) {
             if (item.getName().equalsIgnoreCase(itemName)) {
                 return item;
             }
@@ -87,22 +88,13 @@ public class Config {
         return list;
     }
 
-    private Location parseStringToLoc(String string) {
-        String[] parsedLoc = string.split(",");
-
-        double x = Double.parseDouble(parsedLoc[0]);
-        double y = Double.parseDouble(parsedLoc[1]);
-        double z = Double.parseDouble(parsedLoc[2]);
-
-        return new Location(Bukkit.getWorld("world"), x, y, z);
-    }
-
     private List<Location> getLocationGroundItems() {
         List<Location> locations = new ArrayList<>();
 
         for (String str : config.getStringList("locationGroundItems")) {
-            locations.add(parseStringToLoc(str));
+            locations.add(LocationUtility.parseStringToLoc(config, str));
         }
+
         return locations;
     }
 
