@@ -2,17 +2,21 @@ package fr.cel.essentials.commands.other;
 
 import fr.cel.essentials.Essentials;
 import fr.cel.gameapi.command.AbstractCommand;
+import lombok.Getter;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+
 public class GodCommand extends AbstractCommand {
 
-    private final Essentials main;
+    @Getter private static final List<UUID> playersInGod = new ArrayList<>();
 
-    public GodCommand(Essentials main) {
+    public GodCommand() {
         super("essentials:god", true, true);
-        this.main = main;
     }
 
     @Override
@@ -20,11 +24,11 @@ public class GodCommand extends AbstractCommand {
         Player player = (Player) sender;
         
         if (args.length == 0) {
-            if (main.containsPlayersInGod(player)) {
-                main.getPlayersInGod().remove(player.getUniqueId());
+            if (playersInGod.contains(player.getUniqueId())) {
+                playersInGod.remove(player.getUniqueId());
                 sendMessageWithPrefix(player, "Vous n'êtes plus en mode invulnérable.");
             } else {
-                main.getPlayersInGod().add(player.getUniqueId());
+                playersInGod.add(player.getUniqueId());
                 sendMessageWithPrefix(player, "Vous êtes en mode invulnérable.");
             }
             return;
@@ -33,22 +37,25 @@ public class GodCommand extends AbstractCommand {
         if (args.length == 1) {
             Player target = Bukkit.getPlayer(args[0]);
 
-            if (target != null) {
-                if (main.containsPlayersInGod(target)) {
-                    main.getPlayersInGod().remove(target.getUniqueId());
+            if (isPlayerOnline(target, player)) {
+                if (playersInGod.contains(target.getUniqueId())) {
+                    playersInGod.remove(target.getUniqueId());
                     sendMessageWithPrefix(target, "Vous n'êtes plus en mode invulnérable.");
                     sendMessageWithPrefix(player, "Vous avez enlevé le mode invulnérable de " + target.getName() + ".");
                 } else {
-                    main.getPlayersInGod().add(player.getUniqueId());
+                    playersInGod.add(player.getUniqueId());
                     sendMessageWithPrefix(target, "Vous avez été mis(e) en mode invulnérable.");
                     sendMessageWithPrefix(player, "Vous avez mis en mode invulnérable " + target.getName() + ".");
                 }
-            } else {
-                sendMessageWithPrefix(player, "Ce joueur n'existe pas ou n'est pas connecté.");
             }
 
         }
 
+    }
+
+    @Override
+    protected List<String> onTabComplete(Player player, String[] args) {
+        return null;
     }
 
 }
