@@ -4,6 +4,9 @@ import fr.cel.gameapi.command.AbstractCommand;
 import fr.cel.gameapi.utils.ChatUtility;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+
+import java.util.List;
 
 public class BroadcastCommand extends AbstractCommand {
 
@@ -14,14 +17,17 @@ public class BroadcastCommand extends AbstractCommand {
     @Override
     public void onExecute(CommandSender sender, String[] args) {
         if (args.length >= 1) {
-            final StringBuilder bc = new StringBuilder();
-
+            StringBuilder bc = new StringBuilder();
             for (String part : args) bc.append(part).append(" ");
-
             Bukkit.broadcastMessage(ChatUtility.format("&c[Broadcast]&r&d " + bc));
         } else {
             sendMessageWithPrefix(sender, "La commande est : /broadcast <message>");
         }
+    }
+
+    @Override
+    protected List<String> onTabComplete(Player player, String[] strings) {
+        return null;
     }
 
 }
