@@ -8,6 +8,8 @@ import org.bukkit.entity.Player;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 public class NearCommand extends AbstractCommand {
 
@@ -32,17 +34,23 @@ public class NearCommand extends AbstractCommand {
         if (args.length == 2) {
             Player target = Bukkit.getPlayer(args[1]);
 
-            if (target == null) {
-                sendMessageWithPrefix(player, "Ce joueur n'existe pas ou n'est pas connecté(e).");
-                return;
+            if (isPlayerOnline(target, sender)) {
+                searchPlayers(player, target, args);
             }
-
-            searchPlayers(player, target, args);
         }
     }
 
+    @Override
+    protected List<String> onTabComplete(Player player, String[] args) {
+        if (args.length == 1) {
+            return IntStream.range(1, 10).mapToObj(Integer::toString).collect(Collectors.toList());
+        }
+
+        return null;
+    }
+
     private void searchPlayers(Player player, Player target, String[] args) {
-        int radius = 0;
+        int radius;
         try {
             radius = Integer.parseInt(args[0]);
         } catch (NumberFormatException e) {
