@@ -3,8 +3,7 @@ package fr.cel.hub.inventory;
 import fr.cel.gameapi.GameAPI;
 import fr.cel.gameapi.inventory.AbstractInventory;
 import fr.cel.gameapi.utils.ItemBuilder;
-import fr.cel.hub.inventory.cachecache.CacheCacheInventory;
-import fr.cel.pvp.manager.PVPGameManager;
+import fr.cel.pvp.arena.PVPArenaManager;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
@@ -27,8 +26,8 @@ public class PVPInventory extends AbstractInventory {
     @Override
     public void interact(Player player, String itemName, ItemStack item) {
         switch (item.getType()) {
-            case AMETHYST_BLOCK -> PVPGameManager.getGameManager().getArenaManager().getArenaByDisplayName("Alpha").addPlayer(player);
-            case STONE_BRICKS -> PVPGameManager.getGameManager().getArenaManager().getArenaByDisplayName("Beta").addPlayer(player);
+            case AMETHYST_BLOCK -> PVPArenaManager.getArenaManager().getArenaByDisplayName("Alpha").addPlayer(player);
+            case STONE_BRICKS -> PVPArenaManager.getArenaManager().getArenaByDisplayName("Beta").addPlayer(player);
 
             case BARRIER -> GameAPI.getInstance().getInventoryManager().openInventory(new MinigamesInventory(), player);
 
