@@ -3,7 +3,7 @@ package fr.cel.hub.inventory;
 import fr.cel.gameapi.GameAPI;
 import fr.cel.gameapi.inventory.AbstractInventory;
 import fr.cel.gameapi.utils.ItemBuilder;
-import fr.cel.valocraft.manager.ValoGameManager;
+import fr.cel.valocraft.manager.ValoArenaManager;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
@@ -28,10 +28,10 @@ public class ValocraftInventory extends AbstractInventory {
     @Override
     public void interact(Player player, String itemName, ItemStack item) {
         switch (item.getType()) {
-            case SAND -> ValoGameManager.getGameManager().getArenaManager().getArenaByDisplayName("Désert").addPlayer(player);
-            case SNOW_BLOCK -> ValoGameManager.getGameManager().getArenaManager().getArenaByDisplayName("Neige").addPlayer(player);
-            case CHERRY_LEAVES -> ValoGameManager.getGameManager().getArenaManager().getArenaByDisplayName("Temple").addPlayer(player);
-            case ANDESITE -> ValoGameManager.getGameManager().getArenaManager().getArenaByDisplayName("Complex").addPlayer(player);
+            case SAND -> ValoArenaManager.getArenaManager().getArenaByDisplayName("Désert").addPlayer(player);
+            case SNOW_BLOCK -> ValoArenaManager.getArenaManager().getArenaByDisplayName("Neige").addPlayer(player);
+            case CHERRY_LEAVES -> ValoArenaManager.getArenaManager().getArenaByDisplayName("Temple").addPlayer(player);
+            case ANDESITE -> ValoArenaManager.getArenaManager().getArenaByDisplayName("Complex").addPlayer(player);
 
             case BARRIER -> GameAPI.getInstance().getInventoryManager().openInventory(new MinigamesInventory(), player);
 
