@@ -1,5 +1,6 @@
 package fr.cel.hub.tasks;
 
+import fr.cel.gameapi.utils.LocationUtility;
 import fr.cel.hub.Hub;
 import org.bukkit.*;
 import org.bukkit.configuration.InvalidConfigurationException;
@@ -17,23 +18,20 @@ import java.util.List;
 public class FireworkMusicEvent extends BukkitRunnable {
 
     private final Hub main;
-    private final World world;
     private final List<Location> locations;
 
     public FireworkMusicEvent(Hub main) {
         this.main = main;
-        this.world = Bukkit.getWorld("world");
         this.locations = loadFireworkLocation();
     }
 
     @Override
     public void run() {
         for (Location location : locations) {
-            final Firework firework = (Firework) world.spawnEntity(location, EntityType.FIREWORK);
-            final FireworkMeta fireworkMeta = firework.getFireworkMeta();
+            Firework firework = (Firework) Bukkit.getWorld("world").spawnEntity(location, EntityType.FIREWORK_ROCKET);
+            FireworkMeta fireworkMeta = firework.getFireworkMeta();
 
-            final FireworkEffect fireworkEffect = FireworkEffect.builder().trail(true).withColor(Color.BLUE, Color.WHITE, Color.RED).with(FireworkEffect.Type.STAR).build();
-
+            FireworkEffect fireworkEffect = FireworkEffect.builder().trail(true).withColor(Color.BLUE, Color.WHITE, Color.RED).with(FireworkEffect.Type.STAR).build();
             fireworkMeta.setPower(1);
 
             fireworkMeta.addEffects(fireworkEffect);
@@ -50,7 +48,7 @@ public class FireworkMusicEvent extends BukkitRunnable {
                 config.load(file);
 
                 for (String str : config.getStringList("location")) {
-                    locations1.add(parseStringToLoc(str));
+                    locations1.add(LocationUtility.parseStringToLoc(config, str));
                 }
                 return locations1;
             } catch (IOException | InvalidConfigurationException e) {
@@ -58,16 +56,6 @@ public class FireworkMusicEvent extends BukkitRunnable {
             }
         }
         return null;
-    }
-
-    private Location parseStringToLoc(String string) {
-        String[] parsedLoc = string.split(",");
-
-        double x = Double.parseDouble(parsedLoc[0]);
-        double y = Double.parseDouble(parsedLoc[1]);
-        double z = Double.parseDouble(parsedLoc[2]);
-
-        return new Location(world, x, y, z);
     }
     
 }
