@@ -1,8 +1,9 @@
 package fr.cel.parkour.commands;
 
 import fr.cel.gameapi.command.AbstractCommand;
-import fr.cel.parkour.manager.ParkourGameManager;
-import fr.cel.parkour.manager.area.ParkourMap;
+import fr.cel.gameapi.utils.ChatUtility;
+import fr.cel.parkour.manager.GameManager;
+import fr.cel.parkour.map.ParkourMap;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -13,21 +14,25 @@ import java.util.UUID;
 
 public class ParkourCommands extends AbstractCommand {
 
-    private final ParkourGameManager gameManager;
+    private final GameManager gameManager;
 
-    public ParkourCommands(ParkourGameManager gameManager) {
+    public ParkourCommands(GameManager gameManager) {
         super("parkour:parkour", false, true);
         this.gameManager = gameManager;
     }
 
     @Override
     protected void onExecute(CommandSender sender, String[] args) {
+        if (args.length == 0) {
+            sendHelp(sender);
+            return;
+        }
 
         if (args[0].equalsIgnoreCase("list")) {
-            if (gameManager.getMapManager().getArenas().isEmpty()) {
+            if (gameManager.getMain().getParkourMapManager().getMaps().isEmpty()) {
                 sender.sendMessage(gameManager.getPrefix() + "Aucune map a été installee.");
             } else {
-                gameManager.getMapManager().getArenas().forEach(arena -> sender.sendMessage(gameManager.getPrefix() + "Parkour " + arena.getDisplayName()));
+                gameManager.getMain().getParkourMapManager().getMaps().forEach((s, map) -> sendMessageWithPrefix(sender, "Parkour " + map.getDisplayName()));
             }
             return;
         }
@@ -46,12 +51,12 @@ public class ParkourCommands extends AbstractCommand {
 
         if (args[0].equalsIgnoreCase("listplayer")) {
 
-            if (!gameManager.getMapManager().isPlayerInArena(player)) {
+            if (!gameManager.getMain().getParkourMapManager().isPlayerInArena(player)) {
                 player.sendMessage(gameManager.getPrefix() + "Vous n'êtes pas dans une map Parkour.");
                 return;
             }
 
-            ParkourMap arena = gameManager.getMapManager().getArenaByPlayer(player);
+            ParkourMap arena = gameManager.getMain().getParkourMapManager().getArenaByPlayer(player);
             List<String> playersName = new ArrayList<>();
             for (UUID pls : arena.getPlayers()) {
                 Player player1 = Bukkit.getPlayer(pls);
@@ -59,6 +64,22 @@ public class ParkourCommands extends AbstractCommand {
             }
             player.sendMessage(gameManager.getPrefix() + playersName);
         }
+    }
+
+    @Override
+    protected List<String> onTabComplete(Player player, String[] args) {
+        if (args.length == 1) {
+            return List.of("list", "listplayer", "reload");
+        }
+        return null;
+    }
+
+    private void sendHelp(CommandSender sender) {
+        sender.sendMessage(" ");
+        sender.sendMessage(ChatUtility.format("[Aide pour les commandes du Parkour]", ChatUtility.GOLD));
+        sender.sendMessage("/parkour list : Envoie la liste des maps avec l'état du jeu actuel");
+        sender.sendMessage("/parkour listplayer : Envoie la liste des joueurs dans la partie où vous êtes");
+        sender.sendMessage("/parkour reload : Recharge les cartes");
     }
 
 }
