@@ -1,21 +1,28 @@
 package fr.cel.pvp;
 
 import fr.cel.gameapi.GameAPI;
+import fr.cel.pvp.arena.PVPArenaManager;
+import lombok.Getter;
+import lombok.Setter;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import fr.cel.pvp.commands.PVPCommands;
-import fr.cel.pvp.manager.PVPGameManager;
+import fr.cel.pvp.manager.GameManager;
 
+@Getter
 public class PVP extends JavaPlugin {
 
-    private PVPGameManager gameManager;
+    private GameManager gameManager;
+    @Setter private PVPArenaManager pvpArenaManager;
 
     @Override
     public void onEnable() {
         super.onEnable();
-        this.gameManager = new PVPGameManager(this);
 
-        GameAPI.getInstance().getCommandsManager().addCommand(getCommand("pvp"), new PVPCommands(gameManager));
+        this.gameManager = new GameManager(this);
+        this.pvpArenaManager = new PVPArenaManager(this);
+
+        GameAPI.getInstance().getCommandsManager().addCommand("pvp", new PVPCommands(gameManager), this);
     }
 
     @Override
